new behaviour, based not on name but price, new product field are using (seoText)

+++ app/src/main/java/com/example/iikoapi/startapp/datatype/Nomenclature.kt
@@ -142,7 +142,7 @@ data class Product (
     val groupModifiers: List<GroupModifier>? = null,
     val measureUnit: String? = null,
     val modifiers: List<Modifier>? = null,
-    val price: Long? = null,
+    val price: Int? = null,
 
     @get:JsonProperty("productCategoryId")@field:JsonProperty("productCategoryId")
     val productCategoryID: String? = null,
